+++ devolo_home_control_api/publisher/updater.py
@@ -66,6 +66,12 @@ class Updater:
                         "devolo.WarningBinaryFI:": self._binary_sensor,
                         "hdm": self._device_state}
 
+        unwanted_properties = [".unregistering", "operationStatus"]
+
+        # Early return on unwanted messages
+        if "UNREGISTERED" in message['topic'] or message['properties']['property.name'] in unwanted_properties:
+            return
+
         if "property.name" in message["properties"] \
                 and message['properties']['property.name'] == "pendingOperations" \
                 and "smartGroup" not in message["properties"]["uid"]:
@@ -137,7 +143,7 @@ class Updater:
         element_uid = message['properties']['uid']
 
         # Early return on useless messages
-        if element_uid == "devolo.PairDevice":
+        if element_uid == "devolo.PairDevice" or element_uid == "devolo.RemoveDevice":
             return
 
         pending_operations = bool(message['properties'].get('property.value.new'))
@@ -163,7 +169,8 @@ class Updater:
             return
         if type(message['properties']['property.value.new']) == list \
            and message['properties']['uid'] == "devolo.DevicesPage":
-            self.on_device_change(uids=message['properties']['property.value.new'])
+            device_uid, mode = self.on_device_change(device_uids=message['properties']['property.value.new'])
+            self._publisher.dispatch(device_uid, (device_uid, mode))
 
     def _device_state(self, message: dict):
         """ Update the device state. """
@@ -171,9 +178,11 @@ class Updater:
                         "batteryLow": "battery_low",
                         "status": "status"}
         
-        name = message['properties']['property.name']
         device_uid = message['properties']['uid']
-        value = message['properties']['property.value.new']
+
+        name = message['properties'].get("property.name")
+        value = message['properties'].get("property.value.new")
+
         try:
             self._logger.debug(f"Updating {propery_name[name]} of {device_uid} to {value}")
             setattr(self.devices[device_uid], propery_name[name], value)
@@ -258,7 +267,7 @@ class Updater:
     def _multi_level_switch(self, message: dict):
         """ Update a multi level switch. """
         if not isinstance(message['properties']['property.value.new'], (list, dict, type(None))) \
-                or "smartGroup" not in message["properties"]["uid"]:
+                or "smartGroup" not in message['properties']['uid']:
             element_uid = message['properties']['uid']
             value = message['properties']['property.value.new']
             device_uid = get_device_uid_from_element_uid(element_uid)
@@ -371,6 +380,7 @@ class Updater:
                   "devolo.mprm.gw.GatewayManager",
                   "devolo.mprm.gw.PortalManager",
                   "devolo.smartGroup",
+                  "hdm",
                   "ss",
                   "mcs")
         if not message["properties"]["uid"].startswith(ignore):