+++ docs/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Removed devices are now announced via Publisher
+
+## [v0.15.1] - 2020/10/13
+
+### Fixed
+
+- operationStatus messages were misinterpreted as new state
+
 ## [v0.15.0] - 2020/09/29
 
 ### Added